refactor: 컴포넌트에서 fetch 메서드를 사용하도록 변경
